+++ webpack.config.js
@@ -3,7 +3,7 @@
 const path = require('path');
 
 /**@type {import('webpack').Configuration}*/
-const config = {
+module.exports = {
   target: 'node',
 
   entry: './src/extension.js',
@@ -38,5 +38,3 @@ const config = {
     ],
   },
 };
-
-module.exports = config;